Removed unneeded import. Added new game objects.
Created sections for the various initial items, rooms, player, and droids. Created initial rooms for the ship; still need to add descriptions and hidden items to the rooms. Need to add hidden items to populate rooms. Need to complete the search mechanic. Need to add a score count to track if all the repairs have been made or not.